Overview

Population dynamics across space, in Julia.

Includes high-performance implementations of classic and modern metapopulation models, raster simulations of populations and occurrence, selection on environmental variables, stochasticity of demography and environment, and more.

MetapopulationDynamics.AbstractSpace — Type

AbstractSpace

Abstract type for spatial domains.

MetapopulationDynamics.Raster — Type

Raster{T} <: AbstractSpace where T

A Raster is an AbstractSpace consisting of a grid of sites.

MetapopulationDynamics.SpatialGraph — Type

SpatialGraph{T} <: AbstractSpace

An SpatialGraph is anAbstractSpace` that consists of a set of discrete coordinates.

MetapopulationDynamics.Patches — Type

Patches <: AbstractSpace

TBD

MetapopulationDynamics.numsites — Function

numsites(r::Raster)

Returns the number of nodes in a spatial graph.

numsites(sg::SpatialGraph)

Returns the number of nodes in a spatial graph.

MetapopulationDynamics.distancematrix — Function

distancematrix(r::Raster; distance = Euclidean())

Returns a matrix of pairwise distances for all cells in a Raster

distancematrix(sg::SpatialGraph; distance = Euclidean())

Returns a matrix of pairwise distances for all nodes in a SpatialGraph

Missing docstring.

Missing docstring for Model. Check Documenter's build log for details.

MetapopulationDynamics.AbstractLocalDynamics — Type

abstract type AbstractLocalDynamics <: MetapopulationModel end

Abstract supertype for all models of local dynamics.

MetapopulationDynamics.AbstractOccupancyDynamics — Type

abstract type AbstractOccupancyDynamics <: AbstractLocalDynamics end

MetapopulationDynamics.AbstractAbundanceDynamics — Type

abstract type AbstractAbundanceDynamics <: AbstractLocalDynamics end

MetapopulationDynamics.AbstractDispersalModel — Type

abstract type AbstractDispersalModel <: MetapopulationModel end

Abstract supertype for all dispersal models.

MetapopulationDynamics.modelset — Function

modelset()

Builds a ModelSet.

MetapopulationDynamics.ModelSet — Type

ModelSet{
    L<:Union{AbstractLocalDynamics,Missing},
    D<:Union{AbstractDispersalModel,Missing},
    E<:Union{AbstractEnvironmentModel,Missing},
}

A set of local, environmental, and dispersal models.

MetapopulationDynamics.DispersalKernel — Type

DispersalKernel

A DispersalKernel defines a function that describes the probability of dispersal, with a rate of decay and a threshold, which defines the value at which any result of func is considered to be 0.

MetapopulationDynamics.ExponentialDispersalKernel — Function

ExponentialDispersalKernel

Constructs a DispersalKernel using an exponential function

MetapopulationDynamics.GaussianDispersalKernel — Function

GaussianDispersalKernel

Constructs a DispersalKernel using an Gaussian function

MetapopulationDynamics.DispersalPotential — Type

DispersalPotential

A dispersal potential is a matrix that contains the pairwise probability of dispersal between sites in an AbstractSpace.

Note this is a doubly-stochastic matrix, meaning all rows and columns sum to 0.

MetapopulationDynamics.StochasticJumpDispersalModel — Type

StochasticJumpDispersalModel{T} <: AbstractDispersalModel

MetapopulationDynamics.DeterministicJumpDispersalModel — Type

DeterministicJumpDispersalModel{T} <: AbstractDispersalModel

MetapopulationDynamics.Hanski1994 — Type

Hanski1994{T<:Real} <: AbstractOccupancyDynamics

An incidence-function model of metapopulation dynamics, as defined in Hanski 1994 (citation).

Parameters:

MetapopulationDynamics.Levins1967 — Type

Levins1967{T<:Real} <: AbstractOccupancyDynamics

A Levins metapopulatoin model. Two parameters: c and e.

Citation:

MetapopulationDynamics.RickerModel — Type

RickerModel <: AbstractAbundanceDynamics

A AbstractLocalDyanmics model based on Ricker (citation).

Different sources of stochasticity:

MetapopulationDynamics.StochasticLogistic — Type

StochasticLogistic <: AbstractAbundanceDynamics

The stochastic logistic model, AbstractLocalDynamics.

MetapopulationDynamics.simulate — Function

simulate(
    modelset::M,
    space::S;
    numtimesteps = 100,
) where {M<:ModelSet,S<:AbstractSpace}

Simulation of of ModelSet

simulate(model::M, space::S; numtimesteps = 100) where {M,S<:AbstractSpace}

Simulation of a single Model.

MetapopulationDynamics.simulate! — Function

simulate!(modelset::M, space::S, output) where {M<:ModelSet,S<:AbstractSpace}

Simulation of a ModelSet with an already pre-allocated output.

simulate!(
    model::M,
    space::S,
    output;
    init = rand(Bernoulli(0.5), numsites(space)),
) where {M<:AbstractOccupancyDynamics,S<:AbstractSpace}

Simulation of a single AbstractOccupancyDynamics model.

simulate!(
    model::M,
    space::SpatialGraph,
    output;
    init = rand(Poisson(100), numsites(space)),
) where {M<:AbstractAbundanceDynamics,S}

Simulation of a single AbstractAbundanceDynamics model.

# TODO This should be moved into a different file for dispatch 
specifically on rasters.

MetapopulationDynamics.AbundanceOutput — Type

AbundanceOutput

MetapopulationDynamics.OccupancyOutput — Type

OccupancyOutput

MetapopulationDynamics.pcc — Function

pcc(timeseries)

Computes the mean Pairwise-Crosscorrelation for a timeseries that is an AbundanceOutput.

MetapopulationDynamics.EnvironmentLayer — Type

EnvironmentLayer{S,T}

An environmental layer built on an AbstractSpace.

MetapopulationDynamics.EnvironmentLayerSet — Type

EnvironmentLayerSet

A set of EnvironmentLayers.

Missing docstring.

Missing docstring for AbstractEnvironmentModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for OccupancyEnvironmentModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbundanceEnvironmentModel. Check Documenter's build log for details.

MetapopulationDynamics.numlayers — Function

numlayers(els::EnvironmentLayerSet)

Returns the number of layers in an EnvironmentLayerSet.

MetapopulationDynamics.EnvironmentTimeseries — Type

EnvironmentTimeseries{T<:Union{EnvironmentLayer,EnvironmentLayerSet}}

A time-series of EnvironmentLayers or EnvironmentLayerSets.